Сделал загрузку данных и начал отправку
