+++ .editorconfig
@@ -17,4 +17,4 @@ indent_style = space
 indent_size = 2
 end_of_line = lf
 trim_trailing_whitespace = true
-insert_final_newline = true
+insert_final_newline = false
