+++ js/main.js
@@ -1,12 +1,8 @@
 'use strict';
 
 import './map.js';
-
-
-
-
-
-
+import './get-offers.js';
+import './send-offer.js';
