+++ js/map.js
@@ -1,9 +1,5 @@
 'use strict'
 
-
-import { apartmentsArray } from './data.js';
-import { apartmentsHTML } from './offers.js';
-
 const adForm = document.querySelector('.ad-form');
 const mapFilters = document.querySelector('.map__filters');
 
@@ -31,14 +27,12 @@ const activateApp = () => {
   activateForm(mapFilters);
 }
 
-
 deactivateApp();
-
 
 const mapInitialState = {
   x: 35.685143,
   y: 139.751796,
-  zoom: 13,
+  zoom: 12,
 }
 
 const map = L.map('map-canvas')
@@ -53,25 +47,11 @@ L.tileLayer('https://api.mapbox.com/styles/v1/{id}/tiles/{z}/{x}/{y}?access_toke
   .on('tileerror', deactivateApp)
   .addTo(map);
 
-
 const mainPinIcon = L.icon({
   iconUrl: 'img/main-pin.svg',
   iconSize: 40,
   iconAnchor: [20, 40],
 });
-
-const address = document.querySelector('#address');
-
-L.marker([mapInitialState.x, mapInitialState.y], {
-  icon: mainPinIcon,
-  draggable: true,
-}).on('add', evt => {
-  address.readOnly = true;
-  address.value = `${evt.sourceTarget._latlng.lat.toFixed(5)}, ${evt.sourceTarget._latlng.lng.toFixed(5)}`;
-}).on('drag', evt => {
-  address.value = `${evt.latlng.lat.toFixed(5)}, ${evt.latlng.lng.toFixed(5)}`;
-}).addTo(map);
-
 
 const pinIcon = L.icon({
   iconUrl: 'img/pin.svg',
@@ -79,12 +59,33 @@ const pinIcon = L.icon({
   iconAnchor: [20, 40],
 });
 
-let i = 0;
-apartmentsArray.forEach( apartment => {
-  L.marker([apartment.location.x, apartment.location.y], {
-    icon: pinIcon,
-  }).addTo(map)
-    .bindPopup(
-      apartmentsHTML.querySelectorAll('article')[i]);
-  i++;
+const address = document.querySelector('#address');
+
+const mainPin = L.marker([mapInitialState.x, mapInitialState.y], {
+  icon: mainPinIcon,
+  draggable: true,
+}).on('add', evt => {
+  address.readOnly = true;
+  address.value = `${evt.sourceTarget._latlng.lat.toFixed(5)}, ${evt.sourceTarget._latlng.lng.toFixed(5)}`;
+}).on('drag', evt => {
+  address.value = `${evt.latlng.lat.toFixed(5)}, ${evt.latlng.lng.toFixed(5)}`;
 });
+
+const createPin = pin => {
+  pin.addTo(map);
+};
+
+createPin(mainPin);
+
+export {
+  adForm,
+  mapFilters,
+  blockForm,
+  pinIcon,
+  map,
+  mapInitialState,
+  address,
+  mainPinIcon,
+  mainPin,
+  createPin
+}
